refactor(openmm): extract checkpoint domain into deep module
The checkpoint domain — manifest read/validate, atomic save, orphan
GC, quarantine, terminal classification, production step math —
was scattered across `system_builder.py` (the
`_atomic_save_checkpoint` private function, which was imported by
4 production call sites + 10 test sites despite the underscore
prefix) and `utils.py` (the `load_checkpoint` /
`load_checkpoint_step` / `load_checkpoint_full` trio,
`is_run_complete`, `load_terminal_payload`, `production_ns`).

The seam the underscore prefix advertised did not match the seam
the codebase actually used. Per the v14 architecture review
(improve-codebase-architecture skill, Candidate A), this commit
extracts the domain into a new `biolab_runners.openmm.checkpoint`
deep module with a public interface that matches reality:

  Public interface:
    atomic_save_checkpoint(simulation, output_dir, absolute_step, *, terminal=None)
    quarantine_stale_checkpoint(output_dir)
    load_checkpoint(output_dir) -> LoadedCheckpoint   # was tuple
    is_run_complete(output_dir, config) -> (bool, str)
    load_terminal_payload(output_dir, config) -> dict | None
    production_ns(absolute_step, config) -> float
    InvalidCheckpointError

  LoadedCheckpoint fields:
    absolute_step, state_file_basename, last_record, is_terminal,
    terminal_reason

The internal seam (private helpers used only by this module's
own tests): `_parse_manifest`, `_validate_state_file_reference`,
`_parse_state_filename_step`, `_gc_orphan_states`,
`_check_normal_completion`, `_validate_terminal_payload`,
`_classify_invalid_terminal`, `_production_steps`.

The AGENTS.md invariants around checkpoint / manifest / terminal
/ quarantine / orphan GC / resume safety all describe this module
— they now have a code home that matches.

`system_builder.py` shrinks to a single domain (system / force
field / solvation) at 69% coverage. `utils.py` shrinks to
availability checks + a diagnostic reporter (re-export of
`InvalidCheckpointError` preserved for back-compat).
`runner.py` imports drop from 14 names across two private modules
to 7 names from one public module.

Per DEEPENING.md, old unit tests on shallow modules become waste
once tests at the deepened module's interface exist — the 10
private-import test sites across `test_system_builder.py` and
`test_openmm_runner.py` are deleted, replaced by
`tests/test_checkpoint.py` (40 tests) covering the same
behavior through one public seam.

Test count: 254 passing. Coverage: 81.31% overall,
`checkpoint.py` at 94%.
