+++ biolab_runners/openmm/checkpoint.py
@@ -1,0 +1,948 @@
+"""Checkpoint domain: manifest, state file, atomic save, orphan GC, quarantine, terminal.
+
+This module owns the entire checkpoint lifecycle of an OpenMM MD run:
+
+- :func:`atomic_save_checkpoint` — commits a state file and the manifest
+  in one transaction. The ``os.replace`` on the manifest is the single
+  atomic commit point; a crash before leaves the previous checkpoint
+  active, a crash after leaves the new one active.
+
+- :func:`quarantine_stale_checkpoint` — moves the manifest, energy log,
+  early-abort marker, and every state file into a timestamped
+  ``.stale/<UTC>/`` subdirectory. Called by ``runner.run(force=True)``
+  before building fresh so a stale state cannot be paired with a
+  freshly-built topology.
+
+- :func:`load_checkpoint` — reads the manifest and validates the
+  referenced state file. Raises :class:`InvalidCheckpointError` on
+  any safety gate failure (path traversal, missing file, empty file,
+  step mismatch). Returns a :class:`LoadedCheckpoint` so callers read
+  named fields instead of unpacking tuples.
+
+- :func:`is_run_complete` — the authoritative completion oracle. Tri-state
+  terminal check: a present-but-invalid ``terminal`` payload returns
+  ``(False, "invalid_terminal_<reason>")`` and does NOT fall back to
+  the inferred normal-completion heuristic.
+
+- :func:`load_terminal_payload` — reads and validates the manifest's
+  ``terminal`` field for the early-abort result reconstruction path.
+
+- :func:`production_ns` — the v10 BLOCKER #3 invariant:
+  ``max(0, absolute_step - total_equil_steps) * timestep_fs / 1e6``.
+  Equilibration steps are protocol setup, not scientific progress.
+
+The seam between this module and the runner is the public interface.
+Internal helpers (``_parse_manifest``, ``_validate_state_file_reference``,
+``_parse_state_filename_step``, ``_gc_orphan_states``, ``_check_normal_completion``,
+``_check_manifest_terminal``, ``_production_steps``) are part of the
+module's **internal seam** — they exist so this module's own tests can
+exercise specific failure modes without resorting to setup that goes
+past the public interface. They are NOT exported.
+
+The AGENTS.md invariants around checkpoint / manifest / terminal /
+quarantine / orphan GC / resume safety all describe this module.
+"""
+
+from __future__ import annotations
+
+import json
+import logging
+import os
+import re
+import shutil
+import time
+from dataclasses import dataclass, field
+from datetime import UTC, datetime
+from pathlib import Path
+from typing import TYPE_CHECKING, Any
+
+from biolab_runners.openmm.paths import FileNames
+
+if TYPE_CHECKING:
+    from biolab_runners.openmm.config import OpenMMConfig
+
+logger = logging.getLogger(__name__)
+
+
+# ---------------------------------------------------------------------------
+# Public data shapes
+# ---------------------------------------------------------------------------
+
+
+@dataclass(frozen=True)
+class LoadedCheckpoint:
+    """The parsed manifest plus the validated state-file reference.
+
+    The single entry point for "what is the saved state of this run?"
+    Replaces the previous ``load_checkpoint`` (returns ``(step, file)``),
+    ``load_checkpoint_step`` (returns ``int``), and ``load_checkpoint_full``
+    (returns ``(step, record, file)``) trio — callers read named fields
+    instead of picking the right wrapper.
+
+    Fields:
+        absolute_step: The absolute OpenMM step the saved state is at
+            (the manifest's ``step`` field). Zero means no resumable
+            checkpoint was found.
+        state_file_basename: The basename of the referenced state file
+            (e.g. ``state.42000_12345_170000000.xml``). Empty when no
+            resumable checkpoint exists.
+        last_record: The full last record dict from the manifest, exposed
+            for callers that need to inspect arbitrary fields (the v10
+            ``terminal`` payload is surfaced through the convenience
+            fields below; ``last_record`` is for anything else).
+        is_terminal: True iff ``last_record.terminal`` validated against
+            the v11 schema (``type == "early_abort"``, strict-positive-int
+            ``step`` equal to ``manifest_step``, non-empty ``reason``).
+            A malformed terminal payload leaves ``is_terminal=False``
+            and is surfaced separately via :func:`is_run_complete` —
+            the runner must fail fast on it, not silently treat the run
+            as in-progress or completed.
+        terminal_reason: The human-readable reason string when
+            ``is_terminal`` is True; ``None`` otherwise.
+    """
+
+    absolute_step: int = 0
+    state_file_basename: str = ""
+    last_record: dict[str, Any] = field(default_factory=dict)
+    is_terminal: bool = False
+    terminal_reason: str | None = None
+
+
+class InvalidCheckpointError(Exception):
+    """Raised when ``checkpoint.json`` references a state file that is invalid.
+
+    The runner converts this into a ``result.error`` and refuses to proceed —
+    the user must invoke ``runner.run(force=True)`` to discard the corrupt
+    checkpoint (the quarantine moves the manifest + state files into
+    ``.stale/<UTC>/`` together, so the next non-forced invocation starts from
+    a fresh build).
+    """
+
+
+# ---------------------------------------------------------------------------
+# Pattern constants (internal — used by validation helpers)
+# ---------------------------------------------------------------------------
+
+
+# Pattern for state file basenames accepted by the manifest. The
+# legacy v6 format is a bare ``state.xml``; the v7 format is the
+# generation-versioned ``state.<step>_<pid>_<nanos>.xml``. The
+# validator below rejects anything else (path separators, weird
+# extensions, empty strings) so the resume path cannot be tricked
+# into loading a non-state file or escaping the output directory.
+_STATE_FILENAME_RE = re.compile(r"^state(\.xml|\.\d+_\d+_\d+\.xml)$")
+
+# Pattern for the embedded step in a v7 state filename. The step is
+# the FIRST digit group; the pid and nanos follow. Captured by
+# :func:`_parse_state_filename_step` for the manifest-step ↔
+# state-filename equality check (v10 BLOCKER #1).
+_V7_STATE_STEP_RE = re.compile(r"^state\.(\d+)_\d+_\d+\.xml$")
+
+
+# ---------------------------------------------------------------------------
+# Quarantine — moved from system_builder.py
+# ---------------------------------------------------------------------------
+
+
+# Files that together describe a resumable run. The manifest
+# (``checkpoint.json``) is the source of truth for the saved step
+# and the state file to load. ``energy.csv`` is also kept in the
+# quarantine because it carries the per-step reporter rows and the
+# user may want to inspect a stale trajectory step-by-step.
+#
+# ``early_abort.json`` is the generation-scoped terminal marker
+# written when the offline-mdtraj gate fires (see
+# :meth:`OpenMMRunner._write_abort_metadata`). v9: a forced fresh
+# run must retire this marker together with the rest, otherwise a
+# stale marker from a previous abort run will be re-read by
+# :func:`is_run_complete` and mis-classify a mid-production
+# checkpoint as terminal. The marker is generation-scoped by the
+# manifest step — see the ``force=True quarantine`` rule in AGENTS.md.
+RESUMABLE_FILES: tuple[str, ...] = (
+    FileNames.CHECKPOINT_JSON,
+    FileNames.ENERGY,
+    FileNames.EARLY_ABORT_JSON,
+)
+# Glob pattern for state files. Matches both legacy ``state.xml``
+# (from pre-v7 runs) and the v7 ``state.<step>_<pid>_<nanos>.xml``.
+RESUMABLE_STATE_GLOB = "state*.xml"
+
+
+def quarantine_stale_checkpoint(output_dir: Path) -> list[Path]:
+    """Move resumable files into a timestamped ``.stale/`` subdirectory.
+
+    Used by ``runner.run(force=True)`` to ensure the next non-forced
+    invocation cannot pair a stale state file with a freshly-built
+    topology. The v7 save format uses generation-versioned state
+    files (``state.<step>_<pid>_<nanos>.xml``) referenced by the
+    manifest (``checkpoint.json``) — the manifest is the source of
+    truth for the saved step AND the file to load. So the quarantine
+    must move the manifest, the energy log, AND every state file
+    (both legacy ``state.xml`` and the v7 ``state.<gen>.xml``).
+
+    Returns the list of files actually moved (those that existed).
+    An empty output dir produces an empty list and no ``.stale/``
+    directory is created — there's nothing to quarantine.
+
+    Args:
+        output_dir: Directory holding the stale checkpoint files.
+
+    Returns:
+        List of paths (inside the new ``.stale/<timestamp>/``
+        directory) for the files that were moved.
+    """
+    moved: list[Path] = []
+    # Collect every file that participates in the resume contract:
+    # the manifest, the energy log, and any state file (legacy or
+    # v7 generation-versioned).
+    existing: list[str] = []
+    for name in RESUMABLE_FILES:
+        if (output_dir / name).exists():
+            existing.append(name)
+    for state_file in output_dir.glob(RESUMABLE_STATE_GLOB):
+        existing.append(state_file.name)
+    if not existing:
+        return moved
+
+    # Use UTC + microsecond resolution + PID to avoid filename
+    # collisions on rapid retries (e.g. a CI runner failing and
+    # immediately re-invoking force=True within the same second).
+    # The microsecond + PID combination is unique within a single
+    # host; second-resolution alone was insufficient because two
+    # concurrent invocations in the same second would race the
+    # mkdir(parents=True, exist_ok=False) below.
+    ts = datetime.now(tz=UTC).strftime("%Y%m%dT%H%M%S_%f") + f"_{os.getpid()}"
+    stale_dir = output_dir / ".stale" / ts
+    stale_dir.mkdir(parents=True, exist_ok=False)
+
+    for name in existing:
+        src = output_dir / name
+        dst = stale_dir / name
+        shutil.move(str(src), str(dst))
+        moved.append(dst)
+        logger.info("Quarantined stale checkpoint file: %s -> %s", src, dst)
+
+    return moved
+
+
+# ---------------------------------------------------------------------------
+# Atomic save
+# ---------------------------------------------------------------------------
+
+
+def atomic_save_checkpoint(
+    simulation: object,
+    output_dir: Path,
+    absolute_step: int,
+    *,
+    terminal: dict[str, Any] | None = None,
+) -> str:
+    """Atomically commit a state file plus the manifest as one transaction.
+
+    Design (v7, generation-versioned state files; v10 atomic terminal):
+
+    1. Write a uniquely-named state file directly:
+       ``state.<step>_<pid>_<nanos>.xml``. The filename is the
+       file's identity — there is no canonical ``state.xml``. The
+       filename includes the step and nanosecond timestamp so two
+       concurrent saves from the same process produce distinct
+       files (no collisions).
+
+    2. Write the manifest to a temp file with the new content
+       referencing the state file by basename, then atomically
+       ``os.replace`` it to ``checkpoint.json``. THIS is the
+       single atomic commit point.
+
+       v10 BLOCKER #2: if ``terminal`` is provided, it is embedded
+       into the manifest's last record under a ``terminal`` key.
+       The terminal status therefore commits together with the
+       state file in the same ``os.replace`` — a crash between
+       the two cannot leave a resumable checkpoint whose terminal
+       decision was already made. A separate ``early_abort.json``
+       is written AFTER this function returns (in
+       :meth:`OpenMMRunner._write_abort_metadata`) for downstream
+       consumers — it is a derived file, not authoritative.
+
+    3. Garbage-collect any ``state.*.xml`` that is no longer
+       referenced by the manifest.
+
+    Crash semantics:
+    - Crash before the manifest rename → the previous manifest is
+      still active, the new state file is an orphan (GC'd next run).
+      The resume path loads the previous (coherent) state.
+    - Crash after the manifest rename → the new manifest is active
+      and references the new state file. The resume path loads the
+      new (coherent) state.
+    - Crash mid-``saveState`` → the new state file may be partial
+      or missing, but the manifest still references the previous
+      state file. The resume path loads the previous state.
+
+    The v6 design used two ``os.replace`` calls (one for state.xml,
+    one for the manifest) — those are individually atomic but the
+    pair is not. A crash between the two renames leaves a new state
+    paired with an old manifest; the resume path accepts both
+    because it validates only that the state exists and the manifest
+    has a positive step. The v7 design fixes that by making the
+    manifest rename the single commit point.
+
+    Args:
+        simulation: OpenMM Simulation exposing ``saveState(path)``.
+        output_dir: Directory holding the manifest and the
+            generation-versioned state files.
+        absolute_step: The ABSOLUTE OpenMM step the saved state
+            corresponds to. This is what the runner computed as
+            ``start_step + steps_done`` (where ``start_step`` is
+            the absolute step the simulation was at when the
+            production loop started: ``total_equil_steps`` for
+            fresh runs, ``manifest_step`` for resumed runs). The
+            v6 protocol wrote the invocation-local ``steps_done``
+            instead, which silently broke accounting on resumes.
+        terminal: Optional dict — when provided, embedded into the
+            manifest record under ``terminal``. The caller is
+            responsible for ensuring ``terminal["step"] ==
+            absolute_step`` (enforced by
+            :func:`is_run_complete`).
+
+    Returns:
+        The basename of the saved state file. The runner uses
+        this to populate ``result.state_xml_path``.
+    """
+    pid = os.getpid()
+    nanos = time.time_ns()
+    state_basename = f"state.{absolute_step}_{pid}_{nanos}.xml"
+    state_path = output_dir / state_basename
+
+    # Write the state file directly. The unique filename eliminates
+    # the need for a temp+rename on the state file itself — there
+    # is nothing to overwrite. OpenMM's ``saveState`` writes the
+    # XML and closes the file atomically (per POSIX write+close).
+    simulation.saveState(str(state_path))  # type: ignore[union-attr]
+
+    # Manifest — THIS is the single atomic commit point. The temp
+    # file is written first, then atomically renamed to the
+    # canonical manifest path. If we crash before the rename, the
+    # previous manifest is still active. If we crash after the
+    # rename, the new manifest references the just-written state.
+    manifest_path = output_dir / FileNames.CHECKPOINT_JSON
+    record: dict[str, Any] = {"step": absolute_step, "file": state_basename}
+    if terminal is not None:
+        record["terminal"] = terminal
+    manifest_payload = {"records": [record]}
+    manifest_tmp = manifest_path.with_suffix(manifest_path.suffix + f".tmp.{pid}.{absolute_step}")
+    manifest_tmp.write_text(json.dumps(manifest_payload))
+    os.replace(str(manifest_tmp), str(manifest_path))
+
+    # Garbage-collect orphan state files (best-effort, must not
+    # raise — the save above already succeeded).
+    _gc_orphan_states(output_dir)
+
+    logger.info(
+        "Atomic checkpoint: state=%s manifest=checkpoint.json step=%d terminal=%s",
+        state_basename,
+        absolute_step,
+        "yes" if terminal is not None else "no",
+    )
+    return state_basename
+
+
+def _gc_orphan_states(output_dir: Path) -> None:
+    """Remove ``state.*.xml`` files not referenced by the manifest.
+
+    After each atomic save, the manifest references exactly one
+    state file. Any other ``state.<gen>.xml`` in the directory is
+    an orphan from a previous interrupted save — safe to delete
+    since the resume path would never load it.
+
+    Failures are logged but not raised (orphan cleanup must not
+    interfere with the save that just succeeded).
+
+    Uses :func:`_parse_manifest` (the same internal helper the
+    resume path uses) instead of re-implementing manifest parsing
+    here, so a schema change only needs to be reflected in one
+    place.
+
+    Args:
+        output_dir: Directory holding the manifest and state files.
+    """
+    parsed = _parse_manifest(output_dir)
+    if parsed is None:
+        return
+    _step, _last, active_file = parsed
+    if not active_file:
+        return
+
+    for state_file in output_dir.glob("state.*.xml"):
+        if state_file.name == active_file:
+            continue
+        try:
+            state_file.unlink()
+        except OSError as exc:
+            logger.warning("Could not GC orphan state %s: %s", state_file, exc)
+
+
+# ---------------------------------------------------------------------------
+# Manifest read / validate / load
+# ---------------------------------------------------------------------------
+
+
+def load_checkpoint(output_dir: Path) -> LoadedCheckpoint:
+    """Load the saved step AND state file from the atomic-save manifest.
+
+    The manifest (``checkpoint.json``) is the ONLY authoritative
+    source for the saved state's step and the file to load. It
+    references a state file by basename (e.g.
+    ``state.700000_12345_1700000000000000000.xml``). The resume
+    flow reads both the step and the file from the manifest, then
+    loads the state file via ``simulation.loadState``.
+
+    The previous design saved only ``state.xml`` and inferred the
+    saved step from the last row of ``energy.csv`` — unsafe because
+    the two cadences differ by orders of magnitude (energy.csv at
+    ~10 ps, state.xml at ~2 hr). The v7 fix commits the state file
+    and the manifest as a single atomic transaction: the manifest
+    rename is the only atomic commit point; the state file's
+    filename uniquely identifies it, so any state file not
+    referenced by the manifest is an orphan that can be safely
+    garbage-collected.
+
+    v10 BLOCKER #1: the manifest's ``step`` MUST equal the step
+    encoded in the v7 state filename. A mismatch raises
+    :class:`InvalidCheckpointError`.
+
+    v10 SUGGESTION: malformed manifests (root is a list, records
+    is missing/empty, last record is not a mapping, ``step`` is
+    ``null``, ``file`` is missing) are tolerated as "no resumable
+    checkpoint" — they do NOT crash the runner. The returned
+    :class:`LoadedCheckpoint` carries ``absolute_step=0`` and
+    ``state_file_basename=""`` for that case.
+
+    Args:
+        output_dir: MD output directory.
+
+    Returns:
+        A :class:`LoadedCheckpoint` carrying ``absolute_step``,
+        ``state_file_basename``, ``last_record``, ``is_terminal``,
+        and ``terminal_reason``. The ``is_terminal`` flag is set
+        when ``last_record.terminal`` validates against the v11
+        schema; a malformed terminal payload leaves ``is_terminal``
+        False and is surfaced separately via :func:`is_run_complete`
+        (so the runner fails fast instead of silently treating the
+        run as in-progress or completed).
+
+    Raises:
+        InvalidCheckpointError: If the manifest references a state
+            file that is not a basename, doesn't match the expected
+            pattern, doesn't exist, or is empty. Also raised when
+            the manifest's step doesn't match the step encoded in
+            the v7 state filename (v10 BLOCKER #1). The runner
+            catches this and surfaces it as a user-facing error.
+    """
+    parsed = _parse_manifest(output_dir)
+    if parsed is None:
+        return LoadedCheckpoint()
+    step, last_record, state_file = parsed
+    is_terminal, terminal_reason = _validate_terminal_payload(step, last_record)
+    return LoadedCheckpoint(
+        absolute_step=step,
+        state_file_basename=state_file,
+        last_record=last_record,
+        is_terminal=is_terminal,
+        terminal_reason=terminal_reason,
+    )
+
+
+def _validate_terminal_payload(
+    manifest_step: int, last_record: dict[str, Any]
+) -> tuple[bool, str | None]:
+    """Return ``(is_valid_terminal, reason)`` for the manifest's terminal field.
+
+    ABSENT (no ``terminal`` key) returns ``(False, None)`` — the run
+    is in-progress (or normal-completed, which ``is_run_complete``
+    determines separately).
+
+    A present-but-malformed ``terminal`` returns ``(False, None)`` —
+    the v11 schema is strict (positive-int step equal to
+    ``manifest_step``, ``type == "early_abort"``, non-empty ``reason``).
+    The runner must NOT classify such a run as terminal; it must
+    fail fast via the ``invalid_terminal_*`` reason from
+    :func:`is_run_complete`.
+
+    VALID returns ``(True, "manifest_terminal_<type>_step_<step>")``.
+
+    Args:
+        manifest_step: The step recorded on the manifest's last record.
+        last_record: The last record dict from the manifest.
+
+    Returns:
+        ``(is_valid_terminal, reason)`` — reason is the canonical
+        string used by :func:`is_run_complete` to surface the
+        terminal classification to log lines and the result.
+    """
+    terminal = last_record.get("terminal")
+    if terminal is None:
+        return False, None
+    if not isinstance(terminal, dict):
+        logger.warning(
+            "Manifest has terminal field but it is not a dict; "
+            "treating as invalid terminal payload (NOT as "
+            "normal completion)"
+        )
+        return False, None
+    raw_terminal_step = terminal.get("step")
+    if (
+        not isinstance(raw_terminal_step, int)
+        or isinstance(raw_terminal_step, bool)
+        or raw_terminal_step <= 0
+    ):
+        logger.warning(
+            "Manifest terminal.step is not a positive int "
+            "(got %r); treating as invalid terminal payload",
+            raw_terminal_step,
+        )
+        return False, None
+    if raw_terminal_step != manifest_step:
+        logger.warning(
+            "Manifest has terminal.step=%d but record step=%d; "
+            "treating as invalid terminal payload (NOT as "
+            "normal completion)",
+            raw_terminal_step,
+            manifest_step,
+        )
+        return False, None
+    terminal_type = terminal.get("type")
+    if terminal_type != "early_abort":
+        logger.warning(
+            "Manifest terminal.type=%r is not supported "
+            "(only 'early_abort' is implemented); treating as "
+            "invalid terminal payload",
+            terminal_type,
+        )
+        return False, None
+    reason = terminal.get("reason")
+    if not isinstance(reason, str) or not reason:
+        logger.warning(
+            "Manifest terminal.reason is missing or empty; treating as invalid terminal payload"
+        )
+        return False, None
+    return True, f"manifest_terminal_{terminal_type}_step_{raw_terminal_step}"
+
+
+def _parse_manifest(
+    output_dir: Path,
+) -> tuple[int, dict[str, Any], str] | None:
+    """Internal helper: parse ``checkpoint.json`` with strict validation.
+
+    Returns:
+        ``(step, last_record, state_file)`` on success, ``None`` if
+        the manifest is missing or structurally invalid. The
+        ``last_record`` is the manifest record dict (so callers can
+        inspect the optional ``terminal`` payload).
+
+    Structural validation (v10 SUGGESTION):
+    - Root must be a JSON object (mapping).
+    - ``records`` must be a non-empty list.
+    - Final record must be a JSON object.
+    - ``step`` must be a positive int.
+    - ``file`` must be a non-empty string.
+
+    A structurally invalid manifest returns ``None`` rather than
+    raising — callers (the resume path) treat that as "no
+    resumable checkpoint" and fail fast on orphan state files.
+
+    Raises:
+        InvalidCheckpointError: if the manifest is structurally
+            valid BUT the referenced state file fails the path /
+            existence / size / step-equality checks. The caller
+            catches this and surfaces it as a user-facing error.
+    """
+    manifest_path = output_dir / FileNames.CHECKPOINT_JSON
+    if not manifest_path.exists():
+        return None
+    try:
+        data = json.loads(manifest_path.read_text())
+    except (json.JSONDecodeError, OSError):
+        return None
+
+    # Structural validation (v10 SUGGESTION).
+    if not isinstance(data, dict):
+        return None
+    records = data.get("records")
+    if not isinstance(records, list) or not records:
+        return None
+    last = records[-1]
+    if not isinstance(last, dict):
+        return None
+
+    # Field-type validation — ``step`` must be a positive int and
+    # ``file`` must be a non-empty string.
+    raw_step = last.get("step")
+    raw_file = last.get("file")
+    if not isinstance(raw_step, int) or isinstance(raw_step, bool) or raw_step <= 0:
+        return None
+    if not isinstance(raw_file, str) or not raw_file:
+        return None
+
+    # Path / pattern / existence / size validation. The runner
+    # catches InvalidCheckpointError and surfaces it.
+    _validate_state_file_reference(output_dir, raw_file)
+
+    # v10 BLOCKER #1: manifest step MUST equal the step encoded
+    # in the v7 state filename. Legacy ``state.xml`` has no
+    # embedded step — for legacy compatibility we skip the
+    # equality check (the legacy file can only have been written
+    # by a pre-v7 run; pairing a legacy state with a manifest is
+    # accepted but logged as a compatibility shim).
+    embedded = _parse_state_filename_step(raw_file)
+    if embedded is not None and embedded != raw_step:
+        raise InvalidCheckpointError(
+            f"The manifest step {raw_step} does not match the step "
+            f"encoded in {raw_file!r} ({embedded}). The state file was "
+            f"saved at step {embedded}; pairing it with a manifest step "
+            f"of {raw_step} would silently mismatch the System. Re-run "
+            f"with force=True to discard the checkpoint."
+        )
+    if embedded is None:
+        # Legacy state.xml — no embedded step to cross-validate.
+        # Log a notice so the user can decide to migrate.
+        logger.info(
+            "Manifest references legacy state.xml — the filename carries no "
+            "embedded step, so the manifest step (%d) is trusted as-is. "
+            "Future runs will produce generation-versioned filenames.",
+            raw_step,
+        )
+
+    return raw_step, last, raw_file
+
+
+def _validate_state_file_reference(output_dir: Path, state_file: str) -> Path:
+    """Validate the state file path encoded in a manifest record.
+
+    Returns the absolute path on success. Raises
+    :class:`InvalidCheckpointError` if the reference is unsafe or
+    refers to a file that is missing or empty.
+
+    Validation gates (each must pass):
+
+    1. ``state_file`` is a basename (``Path(s).name == s``).
+       Rejects absolute paths, ``../`` traversal, ``subdir/foo``.
+
+    2. ``state_file`` matches the expected pattern
+       (``state.xml`` or ``state.<digits>_<digits>_<digits>.xml``).
+       Rejects typos, unexpected extensions, anything else.
+
+    3. ``(output_dir / state_file).is_file()`` — the file is on
+       disk. Rejects dangling manifests (the v6 save would have
+       detected this only after ``loadState`` raised).
+
+    4. ``stat().st_size > 0`` — the file is non-empty. Rejects
+       truncated state files (OpenMM's ``saveState`` may write
+       half a file if the process is killed mid-write).
+
+    Args:
+        output_dir: Directory containing the state file.
+        state_file: The basename from the manifest record.
+
+    Returns:
+        The validated absolute path to the state file.
+
+    Raises:
+        InvalidCheckpointError: If any gate fails. The error
+            message includes the offending value and the required
+            remediation (``force=True`` to discard).
+    """
+    if not state_file:
+        raise InvalidCheckpointError(
+            "The manifest references an empty state file path. "
+            "Re-run with force=True to discard the checkpoint."
+        )
+    if Path(state_file).name != state_file:
+        raise InvalidCheckpointError(
+            f"The manifest references a state file with a path "
+            f"that is not a basename: {state_file!r}. Path traversal "
+            f"and absolute paths are rejected. Re-run with force=True "
+            f"to discard the checkpoint."
+        )
+    if not _STATE_FILENAME_RE.match(state_file):
+        raise InvalidCheckpointError(
+            f"The manifest references a state file with an invalid "
+            f"name: {state_file!r}. Expected 'state.xml' (legacy) or "
+            f"'state.<step>_<pid>_<nanos>.xml' (v7). Re-run with "
+            f"force=True to discard the checkpoint."
+        )
+    state_path = output_dir / state_file
+    if not state_path.is_file():
+        raise InvalidCheckpointError(
+            f"The manifest references a state file that does not "
+            f"exist: {state_path}. The run cannot be resumed safely. "
+            f"Re-run with force=True to discard the checkpoint."
+        )
+    if state_path.stat().st_size == 0:
+        raise InvalidCheckpointError(
+            f"The manifest references an empty state file: {state_path}. "
+            f"The state was likely truncated mid-write. Re-run with "
+            f"force=True to discard the checkpoint."
+        )
+    return state_path
+
+
+def _parse_state_filename_step(state_file: str) -> int | None:
+    """Parse the absolute step encoded in a v7 state filename.
+
+    The v7 format is ``state.<step>_<pid>_<nanos>.xml`` — the step
+    is the first digit group. Returns the integer step, or None for
+    the legacy ``state.xml`` (which carries no step encoding and
+    therefore cannot be cross-validated against the manifest's
+    ``step`` field).
+
+    Used by :func:`_parse_manifest` to enforce v10 BLOCKER #1:
+    the manifest's ``step`` MUST equal the step encoded in the v7
+    state filename. A mismatch indicates a corrupt or forged
+    checkpoint and must fail fast.
+
+    Args:
+        state_file: The basename from the manifest record.
+
+    Returns:
+        The embedded step as int, or None if the filename is
+        legacy / unparseable.
+    """
+    m = _V7_STATE_STEP_RE.match(state_file)
+    if m is None:
+        return None
+    return int(m.group(1))
+
+
+# ---------------------------------------------------------------------------
+# Run completion
+# ---------------------------------------------------------------------------
+
+
+def is_run_complete(output_dir: Path, config: OpenMMConfig) -> tuple[bool, str]:
+    """Determine whether a production run is terminal.
+
+    v10 BLOCKER #2: terminal status is now part of the manifest
+    record itself (the ``terminal`` field) and is committed
+    atomically with the state file via the manifest rename — the
+    ``os.replace`` is the single commit point. A separate
+    ``early_abort.json`` may still exist for downstream consumers
+    (``oral_amp.cloud.openmm_cloud``), but it is a derived file
+    written AFTER the atomic save and is NOT consulted by this
+    function for terminal classification.
+
+    A run is terminal when EITHER:
+
+    1. **Normal completion**: ``manifest_step >= total_equil_steps +
+       total_steps``. The manifest's step is the absolute OpenMM
+       step (``start_step + steps_done``), so this signal is
+       unambiguous — not dependent on file sizes or energy row
+       counts.
+
+    2. **Manifest terminal payload**: the manifest's last record
+       contains a ``terminal`` dict with the validated schema
+       (``type == "early_abort"``, ``step == manifest.step`` as a
+       strict positive ``int``, ``reason`` non-empty). A valid
+       payload marks the run terminal; an INVALID payload (present
+       but failing any of these checks) returns
+       ``invalid_terminal_<reason>`` and the runner fails fast
+       with ``force=True`` guidance (v13 BLOCKER). A binding
+       mismatch between ``terminal.step`` and ``manifest.step``
+       indicates data corruption; the runner surfaces this as
+       an error rather than reclassifying the run as in-progress.
+
+    3. **Normal completion** (fallback): when no ``terminal``
+       field is present, ``manifest_step >= total_equil_steps +
+       total_steps`` is treated as normal completion (legacy
+       manifests predate the terminal schema).
+
+    Args:
+        output_dir: MD output directory.
+        config: The OpenMMConfig used to compute total_equil_steps
+            and total_steps.
+
+    Returns:
+        ``(complete, reason)``. ``complete`` is True for the two
+        terminal cases; ``reason`` is a human-readable explanation
+        (e.g. ``"normal_completion_step_50_200_000"``,
+        ``"manifest_terminal_early_abort_step_5_000_000"``,
+        ``"in_progress"``,
+        ``"invalid_terminal_<field>"``).
+
+        v13 BLOCKER: when the manifest has a malformed
+        ``terminal`` payload (the field is present but doesn't
+        validate), the function returns
+        ``(False, "invalid_terminal_<reason>")`` and does NOT fall
+        back to the inferred normal-completion heuristic. The
+        previous behaviour accepted a malformed terminal at the
+        target step as normal completion, contradicting the v11
+        terminal-schema contract.
+
+    Raises:
+        InvalidCheckpointError: If the manifest references a
+            dangling / unsafe / step-mismatched state file. The
+            runner catches this and treats the manifest as
+            unrecoverable (force=True guidance).
+    """
+    checkpoint = load_checkpoint(output_dir)
+    if checkpoint.absolute_step <= 0:
+        return False, "in_progress"
+
+    manifest_step = checkpoint.absolute_step
+    last_record = checkpoint.last_record
+
+    # v13 BLOCKER: tri-state terminal check. Distinguish ABSENT
+    # (no terminal field) from INVALID (terminal field present but
+    # fails schema). The previous implementation collapsed both
+    # into a single ``None`` return, which let a malformed
+    # terminal at the target step fall through to the inferred
+    # normal-completion heuristic — silently reclassifying an
+    # ambiguous/corrupt terminal as a successful run.
+    _is_valid_terminal, terminal_reason = _validate_terminal_payload(manifest_step, last_record)
+    if checkpoint.is_terminal and terminal_reason is not None:
+        # Valid terminal payload. Per v12 BLOCKER precedence,
+        # this takes priority over inferred normal completion,
+        # even when the manifest_step has reached the target.
+        return True, terminal_reason
+    if "terminal" in last_record and terminal_reason is None:
+        # The terminal field is present but malformed. The v11
+        # contract says treat as in-progress with a specific
+        # failure reason. Crucially, do NOT fall back to the
+        # normal-completion heuristic below — the user attempted
+        # to record a terminal decision and got the schema wrong;
+        # the run is in an ambiguous state and the runner must
+        # fail loudly, not silently succeed.
+        invalid_reason = _classify_invalid_terminal(manifest_step, last_record["terminal"])
+        return False, invalid_reason
+    # terminal_state == "absent" — no terminal field; fall back
+    # to the inferred normal completion.
+    normal = _check_normal_completion(manifest_step, config)
+    if normal is not None:
+        return normal
+    return False, "in_progress"
+
+
+def _classify_invalid_terminal(
+    manifest_step: int, terminal: dict[str, object] | str | int | float | bool | None
+) -> str:
+    """Map a malformed ``terminal`` payload to a stable ``invalid_terminal_*`` reason.
+
+    Used by :func:`is_run_complete` to surface the specific schema
+    failure to the runner (and ultimately the user). The reason
+    string is part of the runner's fail-fast error message.
+    """
+    if not isinstance(terminal, dict):
+        return "invalid_terminal_not_dict"
+    raw_step = terminal.get("step")
+    if not isinstance(raw_step, int) or isinstance(raw_step, bool) or raw_step <= 0:
+        return "invalid_terminal_step_invalid_type"
+    if raw_step != manifest_step:
+        return "invalid_terminal_step_mismatch"
+    if terminal.get("type") != "early_abort":
+        return "invalid_terminal_type_unsupported"
+    reason = terminal.get("reason")
+    if not isinstance(reason, str) or not reason:
+        return "invalid_terminal_reason_empty"
+    return "invalid_terminal_unknown"
+
+
+def _check_normal_completion(manifest_step: int, config: OpenMMConfig) -> tuple[bool, str] | None:
+    """Return ``(True, "normal_completion_..._of_<target>")`` if at/past target."""
+    target_step = config.total_equil_steps + config.total_steps
+    if manifest_step >= target_step:
+        return True, f"normal_completion_step_{manifest_step}_of_{target_step}"
+    return None
+
+
+# ---------------------------------------------------------------------------
+# Terminal payload reconstruction
+# ---------------------------------------------------------------------------
+
+
+def load_terminal_payload(output_dir: Path, config: OpenMMConfig) -> dict[str, Any] | None:
+    """Load the validated terminal payload from the manifest.
+
+    v10: the terminal payload is the manifest's ``terminal`` field
+    on the last record. Returns the payload dict only when the
+    run is terminal via the manifest binding (terminal.step ==
+    manifest.step > 0) AND the required fields are well-typed.
+    Returns None otherwise.
+
+    Used by :meth:`OpenMMRunner._reconstruct_terminal_result` to
+    reconstruct the abort result on idempotent reuse.
+
+    Args:
+        output_dir: MD output directory.
+        config: OpenMMConfig (for production_ns computation).
+
+    Returns:
+        The terminal payload dict (with normalised fields
+        ``step``, ``type``, ``reason``, ``production_ns``), or
+        None if the run is not terminal or the payload is
+        invalid.
+
+    Raises:
+        InvalidCheckpointError: If the manifest references a
+            dangling / unsafe / step-mismatched state file.
+    """
+    checkpoint = load_checkpoint(output_dir)
+    if checkpoint.absolute_step <= 0:
+        return None
+    terminal = checkpoint.last_record.get("terminal")
+    if not isinstance(terminal, dict):
+        return None
+    raw_tstep = terminal.get("step")
+    if not isinstance(raw_tstep, int) or isinstance(raw_tstep, bool) or raw_tstep <= 0:
+        return None
+    if raw_tstep != checkpoint.absolute_step:
+        return None
+    # Normalised payload — production_ns is computed from the
+    # v10 BLOCKER #3 invariant, not read from a stored field.
+    payload: dict[str, Any] = {
+        "step": raw_tstep,
+        "type": str(terminal.get("type", "unknown")),
+        "reason": str(terminal.get("reason", "")),
+        "production_ns": production_ns(raw_tstep, config),
+    }
+    # Pass through optional fields if present and well-typed.
+    for opt in ("gate", "target", "peptide_id"):
+        val = terminal.get(opt)
+        if isinstance(val, str):
+            payload[opt] = val
+    return payload
+
+
+# ---------------------------------------------------------------------------
+# Production-time math
+# ---------------------------------------------------------------------------
+
+
+def _production_steps(absolute_step: int, total_equil_steps: int) -> int:
+    """Return the completed production steps for an absolute step.
+
+    Centralises the v10 BLOCKER #3 invariant: every ns reported
+    to downstream consumers (``abort_ns``, ``result.total_ns``,
+    ``md_summary.json``, the reconstructed result) is computed
+    from the COMPLETED PRODUCTION steps (``absolute_step -
+    total_equil_steps``), not the absolute OpenMM step. The
+    equilibration steps are not simulation progress in the
+    scientific sense — they're the protocol's setup.
+
+    Args:
+        absolute_step: The absolute OpenMM step the simulator is
+            at (or the saved state corresponds to).
+        total_equil_steps: The equilibration length in steps
+            (``config.total_equil_steps``).
+
+    Returns:
+        The number of completed production steps (>= 0).
+    """
+    return max(0, absolute_step - total_equil_steps)
+
+
+def production_ns(absolute_step: int, config: OpenMMConfig) -> float:
+    """Convenience: completed production ns for an absolute step.
+
+    Same invariant as :func:`_production_steps`. Returns ns as a
+    float (no rounding) — callers that need a particular precision
+    round at the boundary.
+    """
+    return _production_steps(absolute_step, config.total_equil_steps) * config.timestep_fs / 1e6
